Add support for duplicated items in PriorityQueue

+++ queue/priority_queue.go
@@ -105,12 +105,13 @@ func (items *priorityItems) push(item Item) {
 // items that implement the Item interface and adds them
 // to the queue in priority order.
 type PriorityQueue struct {
-	waiters     waiters
-	items       priorityItems
-	itemMap     map[Item]struct{}
-	lock        sync.Mutex
-	disposeLock sync.Mutex
-	disposed    bool
+	waiters         waiters
+	items           priorityItems
+	itemMap         map[Item]struct{}
+	lock            sync.Mutex
+	disposeLock     sync.Mutex
+	disposed        bool
+	allowDuplicates bool
 }
 
 // Put adds items to the queue.
@@ -127,7 +128,9 @@ func (pq *PriorityQueue) Put(items ...Item) error {
 	}
 
 	for _, item := range items {
-		if _, ok := pq.itemMap[item]; !ok {
+		if pq.allowDuplicates {
+			pq.items.push(item)
+		} else if _, ok := pq.itemMap[item]; !ok {
 			pq.itemMap[item] = struct{}{}
 			pq.items.push(item)
 		}
@@ -186,7 +189,9 @@ func (pq *PriorityQueue) Get(number int) ([]Item, error) {
 		}
 
 		items = pq.items.get(number)
-		deleteItems(items)
+		if !pq.allowDuplicates {
+			deleteItems(items)
+		}
 		sema.response.Done()
 		return items, nil
 	}
@@ -252,9 +257,10 @@ func (pq *PriorityQueue) Dispose() {
 }
 
 // NewPriorityQueue is the constructor for a priority queue.
-func NewPriorityQueue(hint int) *PriorityQueue {
+func NewPriorityQueue(hint int, allowDuplicates bool) *PriorityQueue {
 	return &PriorityQueue{
-		items:   make(priorityItems, 0, hint),
-		itemMap: make(map[Item]struct{}, hint),
+		items:           make(priorityItems, 0, hint),
+		itemMap:         make(map[Item]struct{}, hint),
+		allowDuplicates: allowDuplicates,
 	}
 }
